Finished javascript for new book button
Added event listeners to open and close the modal
to each of the option buttons. I had to add the
type attribute to the "cancel" button because
it kept shooting a submit event. I forgot that by
default, if you make a button without specifying
the type of button, it makes it a submission
button by default, so I gave it an attribute of
button. Issue with the displayAllBooks function
counting through all books, so I added a counter
so that whenever a user creates a new book, the
for loop will start at that book to add it to the
table through DOM manipulation. Also used the
FormData object to parse the data from the
submitted form when the submit event is heard.