Finished library book remove button
I finished the javascript part of the book remove
button next to each book. It adds an event
listener to the remove button so that when
clicked, it searches for the unique book id
from the HTML element in the DOM and removes it
with .remove(). It also removes the book from the
actual book array in the javascript too. I had an
issue with the querySelector not being able to
find some UUID's, but I realized that I assigned
an attribute of "id" not "data-" attribute.
id's cannot start with numbers, so I converted it
into a data- attribute by naming it
"data-book-id" and searched for that one
which worked. I totally forgot to use
data-attributes.