Fixed removing and adding problem
Since I was removing directly from the bookArr,
it changes the index of the books so I had to add
a decrement on the button so that the arrCounter
would be accurate. Also added a utility function
to manually search through the bookArr and remove
the book with the unique ID so that the event
listener doesn't rely on the array index because
it could change.