Finished javascript for read toggle
Read toggle uses the prototype method inside the
Book constructor so not only switch the
read status visually, but also in the array as
well. Made another utility function to
make it so that the button event listener
didn't rely on the for loop index. Also solved
an issue where the toggle button would dissapear,
but that is because rewriting the .textContent
on the table cell overwrites EVERYTHING inside
the cell, so I had to separate the text into
it's own <p> element.