Added empty string detection and whitespace only
Added the constraint to the author and title
input fields where the user cannot submit if
the string is made up of entirely whitespace
characters, or if the user didn't input
anything. I learned that the
.setConstraintValidation changes the error message
AND flags the element as invalid, but it also
needs to be manually turned off if I want to unflag
it as an invalid entry, and if I want to display
it I have to manually define that with .reportValidity.